requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Data Processing and ML
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
torch==2.1.1
yfinance==0.2.31
alpha_vantage==2.3.1
ta-lib==0.4.28
scipy==1.11.3

# Options and Financial Calculations
QuantLib==1.31
py_vollib==1.0.1
pandas_market_calendars==4.3.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Monitoring and Logging
prometheus-client==0.18.0
python-json-logger==2.0.7 